# docs/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: TaskFlow API
  version: 1.0.0
  description: |
    TaskFlow 后端 REST API。
    
    ## 鉴权说明
    
    - 普通用户接口：在请求头中携带 `Authorization: Bearer <JWT>`
    - 管理员控制平面接口（`/admin/*`）：在请求头中携带 `Authorization: Bearer <ADMIN_TOKEN>`
    - 两套鉴权体系完全独立
    
    ## 错误格式
    
    所有错误响应统一格式：
    ```json
    { "error": "错误描述", "code": "ERROR_CODE" }
    ```
    
    ## 时间格式
    
    所有时间字段均为 ISO 8601 格式，UTC 时区，例如 `2025-06-01T12:00:00.000Z`。
servers:
  - url: http://localhost:3001
    description: 本地开发环境
  - url: https://api.taskflow.example.com
    description: 生产环境

# ─── Security Schemes ────────────────────────────────────────────────────────
components:
  securitySchemes:
    userAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 用户 JWT Token，通过 /auth/login 获取
    adminAuth:
      type: http
      scheme: bearer
      description: 管理员 ADMIN_TOKEN，通过环境变量配置

  # ─── Reusable Schemas ──────────────────────────────────────────────────────
  schemas:
    Error:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
          description: 人类可读的错误描述
        code:
          type: string
          description: 机器可读的错误代码
      example:
        error: "资源不存在"
        code: "NOT_FOUND"

    # ── Auth ──
...
    School:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        createdAt:
          type: string
          format: date-time

    # ── Classes ──
...
    AdminUserSummary:
      type: object
      required: [id, email, timezone, isActive, createdAt]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        nickname:
          type: string
          nullable: true
        schoolName:
          type: string
          nullable: true
        studentId:
          type: string
          nullable: true
        timezone:
          type: string
          description: IANA 时区标识符
          example: Asia/Shanghai
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    AdminUpdateUserRequest:
      type: object
      properties:
        nickname:
          type: string
          nullable: true
        isActive:
          type: boolean
        newPassword:
          type: string
          nullable: true
          minLength: 8
    SystemConfigMap:
      type: object
      description: system_config 表的完整键值对，敏感字段值被遮蔽（显示为 ***）
      additionalProperties:
        type: string
      example:
        "app.title": "TaskFlow"
        "app.base_url": "https://taskflow.example.com"
        "auth.registration_open": "true"
        "smtp.host": "smtp.example.com"
        "smtp.password": "***"
        "llm.api_key": "***"
    UpdateSystemConfigRequest:
      type: object
      description: 传入需要更新的键值对，未传入的键不受影响
      additionalProperties:
        type: string
    AdminTestEmailRequest:
      type: object
      required: [to]
      properties:
        to:
          type: string
          format: email
    CreateSchoolRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string

    # ── Pagination ──
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        hasMore:
          type: boolean

# ─── Paths ───────────────────────────────────────────────────────────────────
paths:
...
  /admin/config:
    get:
      tags: [Admin]
      summary: 获取全局系统配置
      description: 敏感字段值显示为 ***。
      security:
        - adminAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemConfigMap'
        '401':
          description: ADMIN_TOKEN 无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      tags: [Admin]
      summary: 更新系统配置
      description: 只更新传入的键，未传入的键不受影响。
      security:
        - adminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSystemConfigRequest'
      responses:
        '200':
          description: 更新成功，返回完整配置
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemConfigMap'
  /admin/config/test-email:
    post:
      tags: [Admin]
      summary: 发送 SMTP 测试邮件
      description: 使用当前系统配置中的 SMTP 参数向指定地址发送测试邮件。
      security:
        - adminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminTestEmailRequest'
      responses:
        '204':
          description: 发送成功
        '400':
          description: SMTP 配置不完整或参数不合法
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /admin/users:
    get:
      tags: [Admin]
      summary: 获取用户列表
      security:
        - adminAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
        - name: q
          in: query
          description: 按邮箱或昵称搜索
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AdminUserSummary'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
  /admin/users/{userId}:
    parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    patch:
      tags: [Admin]
      summary: 修改用户信息 / 停用 / 重置密码
      security:
        - adminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUpdateUserRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserSummary'
        '404':
          description: 用户不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags: [Admin]
      summary: 删除用户
      description: |
        跳过"必须先处理班级"的前置检查，直接执行删除序列。
        该用户拥有的非私有班级不会被自动处理，需管理员另行操作。
      security:
        - adminAuth: []
      responses:
        '204':
          description: 删除成功
        '404':
          description: 用户不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /admin/schools:
    get:
      tags: [Admin]
      summary: 获取学校列表（管理员视图）
      security:
        - adminAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/School'

    post:
      tags: [Admin]
      summary: 添加学校
      security:
        - adminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSchoolRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/School'
        '409':
          description: 学校名称已存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /admin/storage-status:
    get:
      tags: [Admin]
      summary: 获取 S3 存储状态
      description: |
        返回当前 S3 存储后端的连接信息和健康状态。
        用于管理面板展示存储配置和连通性。
      security:
        - adminAuth: []
      responses:
        '200':
          description: 存储状态
          content:
            application/json:
              schema:
                type: object
                properties:
                  endpoint:
                    type: string
                    example: "<account-id>.r2.cloudflarestorage.com"
                  bucket:
                    type: string
                    example: "taskflow-files"
                  useSSL:
                    type: boolean
                  region:
                    type: string
                    example: "auto"
                  connected:
                    type: boolean
                  error:
                    type: string
                    description: 连接失败时的错误信息
                required:
                  - endpoint
                  - bucket
                  - useSSL
                  - region
                  - connected
  /admin/schools/{schoolId}:
    parameters:
      - name: schoolId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    delete:
      tags: [Admin]
      summary: 删除学校
      description: 关联用户和班级的 schoolId 自动置空（FK SET NULL）。
      security:
        - adminAuth: []
      responses:
        '204':
          description: 删除成功
        '404':
          description: 学校不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

# ─── Tags ─────────────────────────────────────────────────────────────────────
...